We should collect the AnsibleJob resources since ACM creates them (#85)
The `AnsibleJob` resources contain valuable status about automation.
Our product shows this status information so I think it's important
we include this in must gather.  Note that I am not attempting to
gather the job logs from the automation.  For now I'm not sure that is
something we want to collect.

Refs:
 - https://github.com/stolostron/backlog/issues/20158

Signed-off-by: Gus Parvin <gparvin@redhat.com>